Compute analysis of variance and likelihood ratio tests of fitted model objects

Compute analysis of variance tables for a fitted model object or a likelihood ratio test for two fitted model objects.

# S3 method for class 'splm'
anova(object, ..., test = TRUE, Terms, L, ddf)

# S3 method for class 'spautor'
anova(object, ..., test = TRUE, Terms, L, ddf)

# S3 method for class 'spglm'
anova(object, ..., test = TRUE, Terms, L, ddf)

# S3 method for class 'spgautor'
anova(object, ..., test = TRUE, Terms, L, ddf)

# S3 method for class 'anova.splm'
tidy(x, ...)

# S3 method for class 'anova.spautor'
tidy(x, ...)

# S3 method for class 'anova.spglm'
tidy(x, ...)

# S3 method for class 'anova.spgautor'
tidy(x, ...)

Arguments

object

A fitted model object from splm(), spautor(), spglm(), or spgautor().

...

An additional fitted model object.

test

A logical value indicating whether p-values from asymptotic Chi-squared hypothesis tests should be returned. Defaults to TRUE.

Terms

An optional character or integer vector that specifies terms in the model used to jointly compute test statistics and p-values (if test = TRUE) against a null hypothesis of zero. Terms is only used when a single fitted model object is passed to the function. If Terms is a character vector, it should contain the names of the fixed effect terms. If Terms is an integer vector, it should correspond to the order (starting at one) of the names of the fixed effect terms. The easiest way to obtain the names of all possible terms is to run tidy(anova(object))$effects (the integer representation matches the positions of this vector).

L

An optional numeric matrix or list specifying linear combinations of the coefficients in the model used to compute test statistics and p-values (if test = TRUE) for coefficient constraints corresponding to a null hypothesis of zero. L is only used when a single fitted model object is passed to the function. If L is a numeric matrix, its rows indicate coefficient constraints and its columns represent coefficients. Then a single hypothesis test is conducted against a null hypothesis of zero. If L is a list, each list element is a numeric matrix specified as above. Then separate hypothesis tests are conducted. The easiest way to obtain all possible coefficients is to run tidy(object)$term.

ddf

The denominator degrees of freedom used. "asymptotic" implements an asymptotic chi-squared test. "satterthwaite" impelments a Satterthwaite/Fai-Cornelius F-test. The default is "satterthwaite" when the sample size is less than or equal to 500 and "asymptotic" otherwise.

x

An object from anova(object).

Value

When ddf is "asymptotic", anova() returns a data frame with degrees of freedom (Df), test statistics (Chi2), and p-values (Pr(>Chi2) if test = TRUE) corresponding to asymptotic Chi-squared hypothesis tests for each model term. When ddf is "satterthwaite", anova() instead numerator degrees of freedom (NumDF), denominator degrees of freedom (DenDF), \(F\) statistics (F value), and p-values (Pr(>F) if test = TRUE) for each model term.

When two fitted model objects are present, anova() returns a data frame with the difference in degrees of freedom between the full and reduced model (Df), a test statistic (Chi2), and a p-value corresponding to the likelihood ratio test (Pr(>Chi2) if test = TRUE).

Whether one or two fitted model objects are provided, tidy() can be used to obtain tidy tibbles of the anova(object) output.

Details

When one fitted model object is present, anova() performs a general linear hypothesis test corresponding to some hypothesis specified by a matrix of constraints. If Terms and L are not specified, each model term is tested against zero (which correspond to type III or marginal hypothesis tests from classical ANOVA). If Terms is specified and L is not specified, all terms are tested jointly against zero. When L is specified, the linear combinations of terms specified by L are jointly tested against zero.

When two fitted model objects are present, one must be a "reduced" model nested in a "full" model. Then anova() performs a likelihood ratio test.

See also

Examples

# one-model anova
spmod <- splm(z ~ water + tarp,
  data = caribou,
  spcov_type = "exponential", xcoord = x, ycoord = y
)
anova(spmod)
#> Analysis of Variance Table
#> 
#> Response: z
#>             NumDF  DenDF F value   Pr(>F)   
#> (Intercept)     1     NA 45.5642       NA   
#> water           1 22.396  1.6738 0.208929   
#> tarp            2 19.525  7.7293 0.003369 **
#> ---
#> Signif. codes:  0 '***' 0.001 '**' 0.01 '*' 0.05 '.' 0.1 ' ' 1
tidy(anova(spmod))
#> # A tibble: 3 × 5
#>   effects     num_df den_df statistic  p.value
#>   <chr>        <int>  <dbl>     <dbl>    <dbl>
#> 1 (Intercept)      1   NA       45.6  NA      
#> 2 water            1   22.4      1.67  0.209  
#> 3 tarp             2   19.5      7.73  0.00337
# see terms
tidy(anova(spmod))$effects
#> [1] "(Intercept)" "water"       "tarp"       
tidy(anova(spmod, Terms = c("water", "tarp")))
#> # A tibble: 1 × 5
#>   effects     num_df den_df statistic p.value
#>   <chr>        <int>  <dbl>     <dbl>   <dbl>
#> 1 water, tarp      3   20.4      5.71 0.00529
# same as
tidy(anova(spmod, Terms = c(2, 3)))
#> # A tibble: 1 × 5
#>   effects     num_df den_df statistic p.value
#>   <chr>        <int>  <dbl>     <dbl>   <dbl>
#> 1 water, tarp      3   20.4      5.71 0.00529
# likelihood ratio test
lmod <- splm(z ~ water + tarp,
  data = caribou,
  spcov_type = "none"
)
tidy(anova(spmod, lmod))
#> # A tibble: 1 × 5
#>   full  reduced    df statistic p.value
#>   <chr> <chr>   <int>     <dbl>   <dbl>
#> 1 spmod lmod        2      2.38   0.304