Glance at a fitted model object

Returns a row of model summaries from a fitted model object. Glance returns the same number of columns for all models and estimation methods. If a particular summary is undefined for a model or estimation method (e.g., likelihood statistics for estimation methods "sv-wls" or "sv-cl" of splm() objects), NA is returned for that summary.

Usage

# S3 method for class 'splm'
glance(x, ...)

# S3 method for class 'spautor'
glance(x, ...)

# S3 method for class 'spglm'
glance(x, ...)

# S3 method for class 'spgautor'
glance(x, ...)

Arguments

x

A fitted model object from splm(), spautor(), spglm(), or spgautor().

...

Other arguments. Not used (needed for generic consistency).

Value

A single-row tibble with columns

n The sample size.

p The number of fixed effects.

npar The number of estimated covariance parameters.

value The optimized value of the fitting function.

AIC The AIC.

AICc The AICc.

BIC The BIC.

logLik The log-likelihood.

deviance The deviance.

pseudo.r.squared The pseudo r-squared.

See also

Examples

spmod <- splm(z ~ water + tarp,
  data = caribou,
  spcov_type = "exponential", xcoord = x, ycoord = y
)
glance(spmod)
#> # A tibble: 1 × 10
#>       n     p  npar value   AIC  AICc   BIC logLik deviance pseudo.r.squared
#>   <int> <dbl> <int> <dbl> <dbl> <dbl> <dbl>  <dbl>    <dbl>            <dbl>
#> 1    30     4     3 -5.85 0.150  1.07  4.35   2.92       26            0.396