Extract fitted model residuals

Extract residuals from a fitted model object. resid is an alias.

# S3 method for class 'splm'
residuals(object, type = "response", ...)

# S3 method for class 'splm'
resid(object, type = "response", ...)

# S3 method for class 'splm'
rstandard(model, ...)

# S3 method for class 'spautor'
residuals(object, type = "response", ...)

# S3 method for class 'spautor'
resid(object, type = "response", ...)

# S3 method for class 'spautor'
rstandard(model, ...)

# S3 method for class 'spglm'
residuals(object, type = "deviance", ...)

# S3 method for class 'spglm'
resid(object, type = "deviance", ...)

# S3 method for class 'spglm'
rstandard(model, ...)

# S3 method for class 'spgautor'
residuals(object, type = "deviance", ...)

# S3 method for class 'spgautor'
resid(object, type = "deviance", ...)

# S3 method for class 'spgautor'
rstandard(model, ...)

Arguments

object

A fitted model object from splm(), spautor(), spglm(), or spgautor().

type

"response" for response residuals, "pearson" for Pearson residuals, or "standardized" for standardized residuals. For splm() and spautor() fitted model objects, the default is "response". For spglm() and spgautor() fitted model objects, deviance residuals are also available ("deviance") and are the default residual type.

...

Other arguments. Not used (needed for generic consistency).

model

A fitted model object from splm(), spautor(), spglm(), or spgautor().

Value

The residuals as a numeric vector.

Details

The response residuals are taken as the response minus the fitted values for the response: \(y - X \hat{\beta}\). The Pearson residuals are the response residuals pre-multiplied by their inverse square root. The standardized residuals are Pearson residuals divided by the square root of one minus the leverage (hat) value. The standardized residuals are often used to check model assumptions, as they have mean zero and variance approximately one.

rstandard() is an alias for residuals(model, type = "standardized").

See also

Examples

spmod <- splm(z ~ water + tarp,
  data = caribou,
  spcov_type = "exponential", xcoord = x, ycoord = y
)
residuals(spmod)
#>            1            2            3            4            5            6 
#>  0.454151895  0.189517422 -0.236909564 -0.084210255  0.040155273  0.177090436 
#>            7            8            9           10           11           12 
#>  0.050789745 -0.162848105 -0.290482578  0.052090436 -0.080848105 -0.123482578 
#>           13           14           15           16           17           18 
#> -0.185909564  0.029155273 -0.136271714 -0.413210255  0.110155273 -0.311271714 
#>           19           20           21           22           23           24 
#> -0.208482578 -0.205909564 -0.015271714  0.029151895 -0.249210255 -0.287844727 
#>           25           26           27           28           29           30 
#> -0.001271714 -0.027210255 -0.286844727 -0.101271714 -0.177848105 -0.190482578 
resid(spmod)
#>            1            2            3            4            5            6 
#>  0.454151895  0.189517422 -0.236909564 -0.084210255  0.040155273  0.177090436 
#>            7            8            9           10           11           12 
#>  0.050789745 -0.162848105 -0.290482578  0.052090436 -0.080848105 -0.123482578 
#>           13           14           15           16           17           18 
#> -0.185909564  0.029155273 -0.136271714 -0.413210255  0.110155273 -0.311271714 
#>           19           20           21           22           23           24 
#> -0.208482578 -0.205909564 -0.015271714  0.029151895 -0.249210255 -0.287844727 
#>           25           26           27           28           29           30 
#> -0.001271714 -0.027210255 -0.286844727 -0.101271714 -0.177848105 -0.190482578 
residuals(spmod, type = "pearson")
#>          1          2          3          4          5          6          7 
#>  2.3706709  1.0054077 -1.2823324 -0.1829072  0.5437284  0.9212618  0.4102518 
#>          8          9         10         11         12         13         14 
#> -0.5841047 -1.2926664  0.7804004 -0.2757715 -0.3606504 -0.5343152  0.8710150 
#>         15         16         17         18         19         20         21 
#> -0.2365538 -2.1127007  1.2989987 -1.1659750 -0.4646992 -0.5333580  0.4386752 
#>         22         23         24         25         26         27         28 
#>  0.8527271 -0.7086751 -0.8987139  0.8066958  0.3958065 -1.0811669  0.2117962 
#>         29         30 
#> -0.2329802 -0.3679197 
residuals(spmod, type = "standardized")
#>          1          2          3          4          5          6          7 
#>  2.5214485  1.0822366 -1.3803208 -0.2012246  0.5907892  0.9978058  0.4465794 
#>          8          9         10         11         12         13         14 
#> -0.6231891 -1.3773902  0.8372453 -0.2922233 -0.3864380 -0.5705954  0.9322631 
#>         15         16         17         18         19         20         21 
#> -0.2520499 -2.2387390  1.3918811 -1.2451453 -0.4973759 -0.5682971  0.4751230 
#>         22         23         24         25         26         27         28 
#>  0.9282355 -0.7560949 -0.9576173  0.8654560  0.4209803 -1.1637850  0.2279805 
#>         29         30 
#> -0.2563122 -0.3997638 
rstandard(spmod)
#>          1          2          3          4          5          6          7 
#>  2.5214485  1.0822366 -1.3803208 -0.2012246  0.5907892  0.9978058  0.4465794 
#>          8          9         10         11         12         13         14 
#> -0.6231891 -1.3773902  0.8372453 -0.2922233 -0.3864380 -0.5705954  0.9322631 
#>         15         16         17         18         19         20         21 
#> -0.2520499 -2.2387390  1.3918811 -1.2451453 -0.4973759 -0.5682971  0.4751230 
#>         22         23         24         25         26         27         28 
#>  0.9282355 -0.7560949 -0.9576173  0.8654560  0.4209803 -1.1637850  0.2279805 
#>         29         30 
#> -0.2563122 -0.3997638