Summarize a fitted model object

Summarize a fitted model object.

# S3 method for class 'splm'
summary(object, ...)

# S3 method for class 'spautor'
summary(object, ...)

# S3 method for class 'spglm'
summary(object, ...)

# S3 method for class 'spgautor'
summary(object, ...)

# S3 method for class 'splmRF'
summary(object, ...)

# S3 method for class 'spautorRF'
summary(object, ...)

Arguments

object

A fitted model object from splm(), spautor(), spglm(), spgautor(), splmRF(), or spautorRF().

...

Other arguments. Not used (needed for generic consistency).

Value

A list with several fitted model quantities used to create informative summaries when printing.

Details

summary() creates a summary of a fitted model object intended to be printed using print(). This summary contains useful information like the original function call, residuals, a coefficients table, a pseudo r-squared, and estimated covariance parameters.

See also

Examples

spmod <- splm(z ~ water + tarp,
  data = caribou,
  spcov_type = "exponential", xcoord = x, ycoord = y
)
summary(spmod)
#> 
#> Call:
#> splm(formula = z ~ water + tarp, data = caribou, spcov_type = "exponential", 
#>     xcoord = x, ycoord = y)
#> 
#> Residuals:
#>      Min       1Q   Median       3Q      Max 
#> -0.41321 -0.20784 -0.11238  0.02915  0.45415 
#> 
#> Coefficients (fixed):
#>             Estimate Std. Error       df t value Pr(>|t|)   
#> (Intercept)  2.05021    0.30373  0.01152   6.750   0.9459   
#> waterY      -0.08336    0.06443 22.39570  -1.294   0.2089   
#> tarpnone     0.08006    0.07750 20.25968   1.033   0.3137   
#> tarpshade    0.28663    0.07657 18.75092   3.743   0.0014 **
#> ---
#> Signif. codes:  0 '***' 0.001 '**' 0.01 '*' 0.05 '.' 0.1 ' ' 1
#> 
#> Pseudo R-squared: 0.3972
#> 
#> Coefficients (exponential spatial covariance):
#>       de       ie    range 
#>  0.10672  0.02244 18.01186 
# \donttest{
sprfmod <- splmRF(log_cond ~ temp + precip, data = lake, spcov_type = "exponential")
summary(sprfmod)
#> ranger:
#> Ranger result
#> 
#> Call:
#>  NA 
#> 
#> Type:                             Regression 
#> Number of trees:                  500 
#> Sample size:                      102 
#> Number of independent variables:  2 
#> Mtry:                             1 
#> Target node size:                 5 
#> Variable importance mode:         none 
#> Splitrule:                        variance 
#> OOB prediction error (MSE):       0.8770362 
#> R squared (OOB):                  0.5129179 
#> 
#> splm on ranger residuals:
#> 
#> Call:
#> NA
#> 
#> Residuals:
#>      Min       1Q   Median       3Q      Max 
#> -2.47041 -0.51053  0.09053  0.67178  3.42486 
#> 
#> Coefficients (fixed):
#>             Estimate Std. Error       df t value Pr(>|t|)
#> (Intercept) -0.08606    0.16793  2.25056  -0.513    0.654
#> 
#> Coefficients (exponential spatial covariance):
#>        de        ie     range 
#> 2.713e-01 6.474e-01 1.238e+05 
# }
# \donttest{
sprfmod <- spautorRF(log_trend ~ stock, data = seal, spcov_type = "car")
summary(sprfmod)
#> ranger:
#> Ranger result
#> 
#> Call:
#>  NA 
#> 
#> Type:                             Regression 
#> Number of trees:                  500 
#> Sample size:                      94 
#> Number of independent variables:  1 
#> Mtry:                             1 
#> Target node size:                 5 
#> Variable importance mode:         none 
#> Splitrule:                        variance 
#> OOB prediction error (MSE):       0.02566385 
#> R squared (OOB):                  0.04627477 
#> 
#> spautor on ranger residuals:
#> 
#> Call:
#> NA
#> 
#> Residuals:
#>      Min       1Q   Median       3Q      Max 
#> -0.40805 -0.09964  0.00282  0.05683  0.82588 
#> 
#> Coefficients (fixed):
#>              Estimate Std. Error        df t value Pr(>|t|)
#> (Intercept)  0.008406   0.018379 29.249014   0.457    0.651
#> 
#> Coefficients (car spatial covariance):
#>      de   range   extra 
#> 0.05101 0.40357 0.02098 
# }