Compute Cook's distance

Compute the Cook's distance for each observation from a fitted model object.

Usage

# S3 method for splm
cooks.distance(model, ...)

# S3 method for spautor
cooks.distance(model, ...)

Arguments

model

A fitted model object from splm() or spautor().

...

Other arguments. Not used (needed for generic consistency).

Value

A vector of Cook's distance values for each observation from the fitted model object.

Details

Cook's distance measures the influence of an observation on a fitted model object. If an observation is influential, its omission from the data noticeably impacts parameter estimates. The larger the Cook's distance, the larger the influence.

See also

Examples

spmod <- splm(z ~ water + tarp,
  data = caribou,
  spcov_type = "exponential", xcoord = x, ycoord = y
)
cooks.distance(spmod)
#>           1           2           3           4           5           6 
#> 0.184522307 0.040365764 0.064916498 0.001740189 0.013456507 0.037035833 
#>           7           8           9          10          11          12 
#> 0.007900142 0.011879854 0.056561044 0.022934810 0.002324453 0.004816503 
#>          13          14          15          16          17          18 
#> 0.010113386 0.027561635 0.001857139 0.136856956 0.062455967 0.047794988 
#>          19          20          21          22          23          24 
#> 0.007927122 0.009617933 0.008341578 0.033395858 0.017414835 0.027307844 
#>          25          26          27          28          29          30 
#> 0.024541392 0.005169888 0.045970784 0.001776621 0.002861920 0.006011727