Heavy metals in mosses near a mining road in Alaska, USA

Heavy metals in mosses near a mining road in Alaska, USA.

Usage

moss

Format

An sf object with 365 rows and 10 columns:

sample: A factor with a sample identifier. Some samples were replicated in the field or laboratory. As a result, there are 318 unique sample identifiers.

field_dup: A factor representing field duplicate. Takes values 1 and 2.

lab_rep: A factor representing laboratory replicate. Takes values 1 and 2.

year: A factor representing year. Takes values 2001 and 2006.

sideroad: A factor representing direction relative to the haul road. Takes values N (north of the haul road) and S (south of the haul road).

log_dist2road: The log of distance (in meters) to the haul road.

log_Zn: The log of zinc concentration in moss tissue (mg/kg).

geometry: POINT geometry representing coordinates in an Alaska Albers projection (EPSG: 3338).

Source

Data were obtained from Peter Neitlich and Linda Hasselbach of the National Park Service. Data were used in the publications listed in References.

References

Neitlich, P.N., Ver Hoef, J.M., Berryman, S. D., Mines, A., Geiser, L.H., Hasselbach, L.M., and Shiel, A. E. 2017. Trends in Spatial Patterns of Heavy Metal Deposition on National Park Service Lands Along the Red Dog Mine Haul Road, Alaska, 2001-2006. PLOS ONE 12(5):e0177936 DOI:10.1371/journal.pone.0177936

Hasselbach, L., Ver Hoef, J.M., Ford, J., Neitlich, P., Berryman, S., Wolk B. and Bohle, T. 2005. Spatial Patterns of Cadmium, Lead and Zinc Deposition on National Park Service Lands in the Vicinity of Red Dog Mine, Alaska. Science of the Total Environment 348: 211-230.