Create a spatial covariance parameter object

Create a spatial covariance parameter object for use with other functions.

spcov_params(spcov_type, de, ie, range, extra, rotate = 0, scale = 1)

Arguments

spcov_type

The spatial covariance function type. Available options include "exponential", "spherical", "gaussian", "triangular", "circular", "cubic", "pentaspherical", "cosine", "wave", "jbessel", "gravity", "rquad", "magnetic", "matern", "cauchy", "pexponential", "car", "sar", and "none".

de

The spatially dependent (correlated) random error variance. Commonly referred to as a partial sill.

ie

The spatially independent (uncorrelated) random error variance. Commonly referred to as a nugget.

range

The correlation parameter.

extra

An extra covariance parameter used when spcov_type is "matern", "cauchy", "pexponential", "car", or "sar".

rotate

Anisotropy rotation parameter (from 0 to \(\pi\) radians). A value of 0 (the default) implies no rotation. Not used if spcov_type is "car" or "sar".

scale

Anisotropy scale parameter (from 0 to 1). A value of 1 (the default) implies no scaling. Not used if spcov_type is "car" or "sar".

Value

A named numeric vector of spatial covariance parameters with class spcov_type.

Details

Generally, all arguments to spcov_params must be specified, though default arguments are often chosen based on spcov_type. When spcov_type is car or sar, ie is assumed to be 0 unless specified otherwise. For full parameterizations of all spatial covariance functions, see spcov_initial().

Examples

spcov_params("exponential", de = 1, ie = 1, range = 1)
#>     de     ie  range rotate  scale 
#>      1      1      1      0      1 
#> attr(,"class")
#> [1] "exponential"