Fit random forest spatial residual models

Fit random forest spatial residual models for point-referenced data (i.e., geostatistical models) using random forest to fit the mean and a spatial linear model to fit the residuals. The spatial linear model fit to the residuals can incorporate variety of estimation methods, allowing for random effects, anisotropy, partition factors, and big data methods.

Usage

splmRF(formula, data, ...)

Arguments

formula

A two-sided linear formula describing the fixed effect structure of the model, with the response to the left of the ~ operator and the terms on the right, separated by + operators.

data

A data frame or sf object object that contains the variables in fixed, random, and partition_factor as well as geographical information. If an sf object is provided with POINT geometries, the x-coordinates and y-coordinates are used directly. If an sf object is provided with POLYGON geometries, the x-coordinates and y-coordinates are taken as the centroids of each polygon.

...

Additional named arguments to ranger::ranger() or splm().

Value

A list with several elements to be used with predict(). These elements include the function call (named call), the random forest object fit to the mean (named ranger), the spatial linear model object fit to the residuals (named spmod or spmod_list), and an object can contain data for locations at which to predict (called newdata). The newdata

object contains the set of observations in data whose response variable is NA. If spcov_type or spcov_initial (which are passed to splm()) are length one, the list has class spmodRF and the spatial linear model object fit to the residuals is called spmod, which has class spmod. If spcov_type or spcov_initial are length greater than one, the list has class spmodRF_list and the spatial linear model object fit to the residuals is called spmod_list, which has class spmod_list. and contains several objects, each with class spmod.

An spmodRF object to be used with predict(). There are three elements: ranger, the output from fitting the mean model with ranger::ranger(); spmod, the output from fitting the spatial linear model to the ranger residuals; and newdata, the newdata

object, if relevant.

Details

The random forest residual spatial linear model is described by Fox Et al. (2020). A random forest model is fit to the mean portion of the model specified by formula using ranger::ranger(). Residuals are computed and used as the response variable in an intercept-only spatial linear model fit using splm(). This model object is intended for use with predict() for perform prediction, also called random forest regression Kriging.

Note

This function does not perform any internal scaling. If optimization is not stable due to large extremely large variances, scale relevant variables so they have variance 1 before optimization.

References

Fox, E.W., Ver Hoef, J. M., & Olsen, A. R. (2020). Comparing spatial regression to random forests for large environmental data sets. PloS one, 15(3), e0229509.

Examples

# \donttest{
sulfate$var <- rnorm(NROW(sulfate)) # add noise variable
sulfate_preds$var <- rnorm(NROW(sulfate_preds)) # add noise variable
sprfmod <- splmRF(sulfate ~ var, data = sulfate, spcov_type = "exponential")
predict(sprfmod, sulfate_preds)
#>          1          2          3          4          5          6          7 
#> 11.3260354 14.4985360  8.8554729 15.4528434  2.7078048 29.0843811 -3.7472819 
#>          8          9         10         11         12         13         14 
#> 15.2077900  6.5062496 17.6143028 -4.2339449 18.8949519  0.8461354 21.7425429 
#>         15         16         17         18         19         20         21 
#> 30.8737497 14.1182966 10.3616990  5.9274988 -3.7144827 13.4781594  5.0286565 
#>         22         23         24         25         26         27         28 
#> -5.4160692 -3.2527590 -4.3117937  4.0895361 -7.5575237 22.8340300 10.2775693 
#>         29         30         31         32         33         34         35 
#> 13.4057889  7.1863819  4.0325676  9.0076379 11.7876560 -1.6011810 -2.7114512 
#>         36         37         38         39         40         41         42 
#>  2.1475384 16.0232487  4.9165419  6.1951623 14.2874082 11.5686687 21.0549982 
#>         43         44         45         46         47         48         49 
#> 14.2518761 30.1271720 -2.8541667  7.1106434 13.6612669  7.8630109 30.5769357 
#>         50         51         52         53         54         55         56 
#>  6.0392232 29.0537832 24.0733400  1.1639907 19.6657160  5.5695725 11.4190362 
#>         57         58         59         60         61         62         63 
#>  8.5224696  3.6418977 22.0808369  2.1956732 10.1888424 18.1241811 -3.0900778 
#>         64         65         66         67         68         69         70 
#> 19.1958757 22.2073973 -0.5849305 -1.9014546  9.6097322  6.5544375 13.7044535 
#>         71         72         73         74         75         76         77 
#>  1.4703563 25.7793427 13.3649952  0.1977996  0.5301885 -2.5390173  8.9823117 
#>         78         79         80         81         82         83         84 
#>  0.7245707 21.3834141 16.3853513 16.7134407 -0.4161622 21.1505571 16.4542252 
#>         85         86         87         88         89         90         91 
#> 10.0369576  5.3034031 -4.8446441  4.2813757  6.9533961 11.9503019 14.5366711 
#>         92         93         94         95         96         97         98 
#> 16.1530883  3.9024402 22.6080707 12.4198031 18.0471193 28.7157686 18.6732068 
#>         99        100 
#> 19.3570959 12.7604056 
# }