Fit random forest spatial residual models

Fit random forest spatial residual models for point-referenced data (i.e., geostatistical models) using random forest to fit the mean and a spatial linear model to fit the residuals. The spatial linear model fit to the residuals can incorporate variety of estimation methods, allowing for random effects, anisotropy, partition factors, and big data methods.

Usage

splmRF(formula, data, ...)

Arguments

formula

A two-sided linear formula describing the fixed effect structure of the model, with the response to the left of the ~ operator and the terms on the right, separated by + operators.

data

A data frame or sf object object that contains the variables in fixed, random, and partition_factor as well as geographical information. If an sf object is provided with POINT geometries, the x-coordinates and y-coordinates are used directly. If an sf object is provided with POLYGON geometries, the x-coordinates and y-coordinates are taken as the centroids of each polygon.

...

Additional named arguments to ranger::ranger() or splm().

Value

A list with several elements to be used with predict(). These elements include the function call (named call), the random forest object fit to the mean (named ranger), the spatial linear model object fit to the residuals (named splm or splm_list), and an object can contain data for locations at which to predict (called newdata). The newdata

object contains the set of observations in data whose response variable is NA. If spcov_type or spcov_initial (which are passed to splm()) are length one, the list has class splmRF and the spatial linear model object fit to the residuals is called splm, which has class splm. If spcov_type or spcov_initial are length greater than one, the list has class splmRF_list and the spatial linear model object fit to the residuals is called splm_list, which has class splm_list. and contains several objects, each with class splm.

An splmRF object to be used with predict(). There are three elements: ranger, the output from fitting the mean model with ranger::ranger(); splm, the output from fitting the spatial linear model to the ranger residuals; and newdata, the newdata

object, if relevant.

Details

The random forest residual spatial linear model is described by Fox et al. (2020). A random forest model is fit to the mean portion of the model specified by formula using ranger::ranger(). Residuals are computed and used as the response variable in an intercept-only spatial linear model fit using splm(). This model object is intended for use with predict() for perform prediction, also called random forest regression Kriging.

Note

This function does not perform any internal scaling. If optimization is not stable due to large extremely large variances, scale relevant variables so they have variance 1 before optimization.

References

Fox, E.W., Ver Hoef, J. M., & Olsen, A. R. (2020). Comparing spatial regression to random forests for large environmental data sets. PloS one, 15(3), e0229509.

Examples

# \donttest{
sulfate$var <- rnorm(NROW(sulfate)) # add noise variable
sulfate_preds$var <- rnorm(NROW(sulfate_preds)) # add noise variable
sprfmod <- splmRF(sulfate ~ var, data = sulfate, spcov_type = "exponential")
predict(sprfmod, sulfate_preds)
#>          1          2          3          4          5          6          7 
#> -4.4851325 32.4148416 18.8629687 11.1070667  0.2634920 44.9187172 10.1639579 
#>          8          9         10         11         12         13         14 
#>  8.5247499  1.8860764 12.4416770 -6.3775122 17.5852110 11.5623177 20.6920841 
#>         15         16         17         18         19         20         21 
#> 21.4332762 11.2546474 12.3473305 12.5740582  7.2474599 28.4453424  0.8100155 
#>         22         23         24         25         26         27         28 
#>  6.0370955 -5.4465345 -6.5956603 -0.5376039  5.5680906 11.5976435 20.3328393 
#>         29         30         31         32         33         34         35 
#> 16.5762511  3.8881088 -4.0676631  6.0720815 14.3105877 -0.7299244  2.2381128 
#>         36         37         38         39         40         41         42 
#>  0.7647218 14.9721270 11.8360909 23.6370043  6.6962919 15.2936410 15.3288531 
#>         43         44         45         46         47         48         49 
#> 22.0592287 25.2815993  3.9245916 19.8398873 11.2798565 11.1203216 14.5816465 
#>         50         51         52         53         54         55         56 
#> -5.3944058 31.9331112  9.0964024  9.8282052 -1.1915019  8.2903132 16.1091002 
#>         57         58         59         60         61         62         63 
#> 22.0095733 -9.3955489 13.9137869 14.7992536 13.2129057 17.4727237  8.3532127 
#>         64         65         66         67         68         69         70 
#> 17.7958098 27.1275193 -1.4759576 -4.5544598  6.2849949  5.3399282  9.4943331 
#>         71         72         73         74         75         76         77 
#> -3.3325025 24.7863430  1.4071574  6.0413836  2.3921251 -0.6820872 21.0012823 
#>         78         79         80         81         82         83         84 
#>  6.1881032 16.7233182 12.3833482  9.2240733 12.0010120  8.2314757  6.5824609 
#>         85         86         87         88         89         90         91 
#>  5.1828613  8.7499919  4.8092601 -3.0923806  9.9381419 17.9122543 15.8968522 
#>         92         93         94         95         96         97         98 
#> 10.5204847  4.9098156 20.0045755  9.4817227 20.2961642 20.3496114 19.8505763 
#>         99        100 
#> 17.7304013 10.3742592 
# }