spmodel: Spatial Statistical Modeling and Prediction

Fit, summarize, and predict for a variety of spatial statistical models applied to point-referenced and areal (lattice) data. Parameters are estimated using various methods. Additional modeling features include anisotropy, non-spatial random effects, partition factors, big data approaches, and more. Model-fit statistics are used to summarize, visualize, and compare models. Predictions at unobserved locations are readily obtainable. For additional details, see Dumelle et al. (2023) doi:10.1371/journal.pone.0282524 .

See also

Author

Maintainer: Michael Dumelle Dumelle.Michael@epa.gov (ORCID)

Authors:

Matt Higham mhigham@stlawu.edu (ORCID)

Jay M. Ver Hoef jay.verhoef@noaa.gov (ORCID)

Other contributors:

Ryan A. Hill hill.ryan@epa.gov (ORCID) [contributor]