Simulate a spatial normal (Gaussian) random variable

Simulate a spatial normal (Gaussian) random variable with a specific mean and covariance structure.

Usage

sprnorm(
  spcov_params,
  mean = 0,
  samples = 1,
  data,
  randcov_params,
  partition_factor,
  ...
)

# S3 method for exponential
sprnorm(
  spcov_params,
  mean = 0,
  samples = 1,
  data,
  randcov_params,
  partition_factor,
  xcoord,
  ycoord,
  ...
)

# S3 method for none
sprnorm(
  spcov_params,
  mean = 0,
  samples = 1,
  data,
  randcov_params,
  partition_factor,
  ...
)

# S3 method for car
sprnorm(
  spcov_params,
  mean = 0,
  samples = 1,
  data,
  randcov_params,
  partition_factor,
  W,
  row_st = TRUE,
  M,
  ...
)

Arguments

spcov_params

An spcov_params() object.

mean

A numeric vector representing the mean. mean must have length 1 (in which case it is recycled) or length equal to the number of rows in data. The default is 0.

samples

The number of independent samples to generate. The default is 1.

data

A data frame or sf object containing spatial information.

randcov_params

A randcov_params() object.

partition_factor

A formula indicating the partition factor.

...

Other arguments. Not used (needed for generic consistency).

xcoord

Name of the column in data representing the x-coordinate. Can be quoted or unquoted. Not required if data are an sf object.

ycoord

Name of the column in data representing the y-coordinate. Can be quoted or unquoted. Not required if data are an sf object.

W

Weight matrix specifying the neighboring structure used for car and sar models. Not required if data are an sf polygon object and W should be calculated internally.

row_st

A logical indicating whether row standardization be performed on W. The default is TRUE.

M

M matrix satisfying the car symmetry condition. The car symmetry condition states that \((I - range * W)^{-1}M\) is symmetric, where \(I\) is an identity matrix, \(range\) is a constant that controls the spatial dependence, W is the weights matrix, and \(^{-1}\) represents the inverse operator. M is required for car models when W is provided and row_st is FALSE. When M, is required, the default is the identity matrix.

Value

If samples is 1, a vector of random variables for each row of data

is returned. If samples is greater than one, a matrix of random variables is returned, where the rows correspond to each row of data and the columns correspond to independent samples.

Details

Random variables are simulated via the product of the covariance matrix's square (Cholesky) root and independent standard normal random variables with mean 0 and variance 1. Computing the square root is a significant computational burden and likely unfeasible for sample sizes much past 10,000. Because this square root only needs to be computed once, however, it is nearly the sample computational cost to call sprnorm() for any value of samples.

Only methods for the exponential, none, and car covariance functions are documented here, but methods exist for all other spatial covariance functions defined in spcov_initial(). Syntax for the exponential method is the same as syntax for spherical, gaussian, triangular, circular, cubic, pentaspherical, cosine, wave, jbessel, gravity, rquad, magnetic, matern, cauchy, and pexponential methods. Syntax for the car method is the same as syntax for the sar method. The extra parameter for car and sar models is ignored when all observations have neighbors.

Examples

spcov_params_val <- spcov_params("exponential", de = 1, ie = 1, range = 1)
sprnorm(spcov_params_val, data = caribou, xcoord = x, ycoord = y)
#>  [1] -0.473964281 -1.200513615 -0.474099910 -1.793888707 -0.074168856
#>  [6] -1.508407632 -1.211535861 -1.924602152 -0.193508976 -0.002052277
#> [11]  0.063011180  0.344156349 -0.948780615 -0.350705744 -0.135785096
#> [16]  0.519994308  0.689117752 -2.001440378  0.972466141 -0.792149822
#> [21]  1.080118804  0.953985833  0.554521457  0.178454735  2.691408694
#> [26]  1.269644278  3.719608898 -0.019729351 -0.230618010  1.753592644
sprnorm(spcov_params_val, mean = 1:30, samples = 5, data = caribou, xcoord = x, ycoord = y)
#>            [,1]      [,2]      [,3]      [,4]       [,5]
#>  [1,]  1.959814  2.361479  3.308520  3.305414  1.1401848
#>  [2,]  4.705892  2.028682  3.797394  2.403564  0.7944248
#>  [3,]  5.668380  3.247860  2.735391  2.317678  3.0010535
#>  [4,]  6.201168  2.875584  4.450310  3.311665  5.3180126
#>  [5,]  4.929344  5.601934  5.750809  7.620151  6.7010850
#>  [6,]  8.100227  3.909230  7.348376  8.032601  5.3526803
#>  [7,]  6.119970  6.532758  9.715175  7.385059  9.2246052
#>  [8,]  8.331121  5.955490  7.177578  7.773055  7.0652128
#>  [9,]  6.432769  9.617972  9.624435  9.692302 11.1874696
#> [10,] 11.857431  9.703247  8.901158 10.483929 11.9609324
#> [11,]  9.936470 13.472820 10.252977  9.158477 10.6531257
#> [12,] 12.714734 12.429776 13.707480 10.040952 11.8707936
#> [13,] 10.007339 13.309016 10.873372 13.023500 12.6839842
#> [14,] 15.502477 12.842745 14.545008 12.905959 14.4075815
#> [15,] 15.797439 15.474844 17.790912 15.150106 15.7532839
#> [16,] 16.169045 14.989811 14.970861 14.332831 15.9684165
#> [17,] 15.995921 18.578402 16.236862 17.415591 13.6446743
#> [18,] 19.327413 19.559180 16.137012 16.404407 18.5628867
#> [19,] 19.659653 18.957724 16.675515 19.859471 20.4956737
#> [20,] 19.790992 21.374007 20.973778 19.818530 19.1899627
#> [21,] 23.663944 19.828936 18.114615 23.033196 19.2073410
#> [22,] 21.510705 21.762511 21.473857 20.784469 20.7311076
#> [23,] 22.284473 24.106970 20.757314 21.170187 21.9569545
#> [24,] 24.265006 23.562653 24.341100 23.678379 22.0388071
#> [25,] 24.162281 27.092215 26.051400 21.741349 25.6761491
#> [26,] 26.275020 25.108990 25.406309 24.613334 26.4534224
#> [27,] 27.417955 26.135045 27.724811 24.230693 27.3768537
#> [28,] 27.520851 29.906723 29.010994 25.785666 27.1210973
#> [29,] 28.927303 29.892832 30.445955 28.424295 29.8437689
#> [30,] 29.870487 29.358744 28.671426 30.591777 31.6219468