Simulate a spatial normal (Gaussian) random variable

Simulate a spatial normal (Gaussian) random variable with a specific mean and covariance structure.

Usage

sprnorm(
  spcov_params,
  mean = 0,
  samples = 1,
  data,
  randcov_params,
  partition_factor,
  ...
)

# S3 method for exponential
sprnorm(
  spcov_params,
  mean = 0,
  samples = 1,
  data,
  randcov_params,
  partition_factor,
  xcoord,
  ycoord,
  ...
)

# S3 method for none
sprnorm(
  spcov_params,
  mean = 0,
  samples = 1,
  data,
  randcov_params,
  partition_factor,
  ...
)

# S3 method for car
sprnorm(
  spcov_params,
  mean = 0,
  samples = 1,
  data,
  randcov_params,
  partition_factor,
  W,
  row_st = TRUE,
  M,
  ...
)

Arguments

spcov_params

An spcov_params() object.

mean

A numeric vector representing the mean. mean must have length 1 (in which case it is recycled) or length equal to the number of rows in data. The default is 0.

samples

The number of independent samples to generate. The default is 1.

data

A data frame or sf object containing spatial information.

randcov_params

A randcov_params() object.

partition_factor

A formula indicating the partition factor.

...

Other arguments. Not used (needed for generic consistency).

xcoord

Name of the column in data representing the x-coordinate. Can be quoted or unquoted. Not required if data are an sf object.

ycoord

Name of the column in data representing the y-coordinate. Can be quoted or unquoted. Not required if data are an sf object.

W

Weight matrix specifying the neighboring structure used for car and sar models. Not required if data are an sf polygon object and W should be calculated internally.

row_st

A logical indicating whether row standardization be performed on W. The default is TRUE.

M

M matrix satisfying the car symmetry condition. The car symmetry condition states that \((I - range * W)^{-1}M\) is symmetric, where \(I\) is an identity matrix, \(range\) is a constant that controls the spatial dependence, W is the weights matrix, and \(^{-1}\) represents the inverse operator. M is required for car models when W is provided and row_st is FALSE. When M, is required, the default is the identity matrix.

Value

If samples is 1, a vector of random variables for each row of data

is returned. If samples is greater than one, a matrix of random variables is returned, where the rows correspond to each row of data and the columns correspond to independent samples.

Details

Random variables are simulated via the product of the covariance matrix's square (Cholesky) root and independent standard normal random variables with mean 0 and variance 1. Computing the square root is a significant computational burden and likely unfeasible for sample sizes much past 10,000. Because this square root only needs to be computed once, however, it is nearly the sample computational cost to call sprnorm() for any value of samples.

Only methods for the exponential, none, and car covariance functions are documented here, but methods exist for all other spatial covariance functions defined in spcov_initial(). Syntax for the exponential method is the same as syntax for spherical, gaussian, triangular, circular, cubic, pentaspherical, cosine, wave, jbessel, gravity, rquad, magnetic, matern, cauchy, and pexponential methods. Syntax for the car method is the same as syntax for the sar method. The extra parameter for car and sar models is ignored when all observations have neighbors.

Examples

spcov_params_val <- spcov_params("exponential", de = 1, ie = 1, range = 1)
sprnorm(spcov_params_val, data = caribou, xcoord = x, ycoord = y)
#>  [1] -0.4141798947  0.8754033261  0.9909964440 -1.6854436981 -0.7177066662
#>  [6]  0.7221768068  1.0608685327 -0.4455809628 -0.0358468581 -1.2331012441
#> [11] -0.0121050047  0.6825719401  1.8885684238  3.2357646562 -0.3236846424
#> [16]  0.0008987733  0.9830467100  1.7653545195  3.2190882541  0.4528468185
#> [21] -1.1291504579 -0.5011889270  1.1625324029 -2.3406347555 -0.3694399139
#> [26]  0.2055746001  0.2298870777  2.8240301690  1.6138756655 -1.6500186891
sprnorm(spcov_params_val, mean = 1:30, samples = 5, data = caribou, xcoord = x, ycoord = y)
#>            [,1]      [,2]       [,3]       [,4]       [,5]
#>  [1,]  2.229153  1.462321  2.2411053 -0.1408958  0.2325348
#>  [2,]  4.529805  2.672669  0.5342035  3.2138382  1.6346942
#>  [3,]  3.650417  2.045490  2.2851819  3.2086407  1.9227771
#>  [4,]  3.534750  6.396168  4.0019210  4.0071680  4.4084278
#>  [5,]  6.966512  7.146783  3.3932743  6.9847530  4.6629422
#>  [6,]  8.380621  6.742105  6.5885822  3.9309706  6.0593199
#>  [7,]  8.053691  7.309490  6.8026510  4.8531421  7.2892191
#>  [8,]  9.086599  8.052997  7.1165832  8.4161032  7.8817427
#>  [9,] 12.508193  9.388723  8.3195101  8.9181296 10.2966378
#> [10,] 12.614975 11.788532 13.4153772 11.4733615  9.4826048
#> [11,] 10.270888  9.830519 11.5702429 10.9913194 10.7931955
#> [12,] 10.292714 12.372273 11.5235581 11.1753656 12.4044837
#> [13,] 12.019957 11.354008 10.0455471 14.9210061 13.6580889
#> [14,] 14.732081 14.392766 12.7755348 14.7802836 14.9542733
#> [15,] 15.730486 16.094785 15.7316027 15.4354447 16.1953193
#> [16,] 14.462285 13.331054 18.1387870 17.0072061 15.4240054
#> [17,] 17.123930 16.974889 17.2568705 17.7215753 14.7114688
#> [18,] 14.790494 18.594710 15.9820744 17.5011687 17.9936770
#> [19,] 20.612240 21.739544 16.6765675 21.0562899 20.2412626
#> [20,] 22.441689 21.982433 18.5820632 23.1674380 21.0176501
#> [21,] 20.964724 21.175957 21.9097190 22.8648442 22.7448473
#> [22,] 20.929465 22.381041 21.8282337 20.5903964 21.5601355
#> [23,] 22.503227 24.099690 23.8567381 21.9887783 22.5044018
#> [24,] 25.853218 25.098367 24.8439059 23.8845485 25.1315176
#> [25,] 27.326652 25.147286 25.8221885 25.1024889 25.7186829
#> [26,] 24.142632 25.089201 29.3006111 25.6247783 27.1157621
#> [27,] 27.722113 28.097002 29.4875980 27.2047036 25.7655944
#> [28,] 29.460726 29.891445 27.8690744 25.6065189 30.1559171
#> [29,] 29.009647 29.356898 26.1648700 29.5369171 30.0284159
#> [30,] 33.147983 29.479773 30.8956972 28.4501954 29.4671703