Simulate a spatial normal (Gaussian) random variable

Simulate a spatial normal (Gaussian) random variable with a specific mean and covariance structure.

Usage

sprnorm(
  spcov_params,
  mean = 0,
  samples = 1,
  data,
  randcov_params,
  partition_factor,
  ...
)

# S3 method for exponential
sprnorm(
  spcov_params,
  mean = 0,
  samples = 1,
  data,
  randcov_params,
  partition_factor,
  xcoord,
  ycoord,
  ...
)

# S3 method for none
sprnorm(
  spcov_params,
  mean = 0,
  samples = 1,
  data,
  randcov_params,
  partition_factor,
  ...
)

# S3 method for car
sprnorm(
  spcov_params,
  mean = 0,
  samples = 1,
  data,
  randcov_params,
  partition_factor,
  W,
  row_st = TRUE,
  M,
  ...
)

Arguments

spcov_params

An spcov_params() object.

mean

A numeric vector representing the mean. mean must have length 1 (in which case it is recycled) or length equal to the number of rows in data. The default is 0.

samples

The number of independent samples to generate. The default is 1.

data

A data frame or sf object containing spatial information.

randcov_params

A randcov_params() object.

partition_factor

A formula indicating the partition factor.

...

Other arguments. Not used (needed for generic consistency).

xcoord

Name of the column in data representing the x-coordinate. Can be quoted or unquoted. Not required if data are an sf object.

ycoord

Name of the column in data representing the y-coordinate. Can be quoted or unquoted. Not required if data are an sf object.

W

Weight matrix specifying the neighboring structure used for car and sar models. Not required if data are an sf polygon object and W should be calculated internally.

row_st

A logical indicating whether row standardization be performed on W. The default is TRUE.

M

M matrix satisfying the car symmetry condition. The car symmetry condition states that \((I - range * W)^{-1}M\) is symmetric, where \(I\) is an identity matrix, \(range\) is a constant that controls the spatial dependence, W is the weights matrix, and \(^{-1}\) represents the inverse operator. M is required for car models when W is provided and row_st is FALSE. When M, is required, the default is the identity matrix.

Value

If samples is 1, a vector of random variables for each row of data

is returned. If samples is greater than one, a matrix of random variables is returned, where the rows correspond to each row of data and the columns correspond to independent samples.

Details

Random variables are simulated via the product of the covariance matrix's square (Cholesky) root and independent standard normal random variables with mean 0 and variance 1. Computing the square root is a significant computational burden and likely unfeasible for sample sizes much past 10,000. Because this square root only needs to be computed once, however, it is nearly the sample computational cost to call sprnorm() for any value of samples.

Only methods for the exponential, none, and car covariance functions are documented here, but methods exist for all other spatial covariance functions defined in spcov_initial(). Syntax for the exponential method is the same as syntax for spherical, gaussian, triangular, circular, cubic, pentaspherical, cosine, wave, jbessel, gravity, rquad, magnetic, matern, cauchy, and pexponential methods. Syntax for the car method is the same as syntax for the sar method. The extra parameter for car and sar models is ignored when all observations have neighbors.

Examples

spcov_params_val <- spcov_params("exponential", de = 1, ie = 1, range = 1)
sprnorm(spcov_params_val, data = caribou, xcoord = x, ycoord = y)
#>  [1]  0.3912812  0.8777097 -4.1302048  2.2797164 -0.4956502  1.6418570
#>  [7] -1.8932360  0.5352149  2.0619029 -0.9534150 -0.5667603 -2.5093976
#> [13]  1.8618835  2.1479148 -1.3046371 -3.1207777  0.9716937  0.9996079
#> [19]  2.5964685  0.7714237 -0.9661860  0.1189094  2.0509624  0.4091457
#> [25] -1.7011496  0.8097515  2.0975446  0.9428908 -0.3261922  0.7745804
sprnorm(spcov_params_val, mean = 1:30, samples = 5, data = caribou, xcoord = x, ycoord = y)
#>            [,1]      [,2]       [,3]        [,4]       [,5]
#>  [1,]  1.872950  3.814811  1.0384786  0.34103816  0.2681738
#>  [2,]  2.189090  4.323944 -0.6117392  0.01048705  2.8181024
#>  [3,]  4.290003  2.957607 -0.6908909  1.67251457  2.8266280
#>  [4,]  5.340555  5.852864  3.4485193  2.94876572  3.9896520
#>  [5,]  5.563930  4.965418  3.1888953  4.48009971  3.1712655
#>  [6,]  4.913318  7.720977  5.2389323  5.31166662  8.0997392
#>  [7,]  8.058354 10.137098  5.2305738  6.85291246  6.7435056
#>  [8,]  9.535089  8.115062  7.1943304  6.51618319  7.3935217
#>  [9,] 10.667224 10.149462  9.2787007  7.53171164  7.6061350
#> [10,] 12.148142 10.419532  9.6816257  9.61463656 10.4043618
#> [11,] 10.250098 12.819001 10.1724534  9.81954247 11.2505413
#> [12,] 13.289633  9.902156  9.5707018 10.36823739 13.0977571
#> [13,] 15.171664 12.858070 10.7438697 12.26694784 14.7155336
#> [14,] 15.738534 13.364659 16.8171465 13.52160994 12.3949796
#> [15,] 17.698352 15.880055 16.8796074 14.47011690 14.1415197
#> [16,] 16.370792 19.142706 17.0077548 18.28717768 15.8523331
#> [17,] 17.582048 16.350445 15.6317371 15.51354726 16.0702756
#> [18,] 19.584679 16.396006 18.5562000 15.65432319 16.7338999
#> [19,] 20.390726 17.463722 18.2159034 18.12525404 18.1196701
#> [20,] 20.337776 19.564041 22.1216450 19.80567868 20.7816745
#> [21,] 22.126702 20.676012 20.8479439 20.36265356 23.2925062
#> [22,] 22.304049 23.102541 22.6000395 20.90305138 22.9828599
#> [23,] 23.463388 20.671360 22.8633382 20.53315107 21.3934476
#> [24,] 23.997335 23.190577 23.3996608 22.74044050 22.9416516
#> [25,] 26.277090 24.289272 23.6681108 24.43870779 23.2708006
#> [26,] 27.271521 25.266712 25.4925351 25.54047728 29.7606836
#> [27,] 27.088933 25.967296 25.0684743 26.74042523 29.0202116
#> [28,] 28.818220 27.171454 27.8196242 28.90718431 28.0870971
#> [29,] 29.499937 27.167490 27.3813456 30.20046857 28.2552536
#> [30,] 29.001238 29.861259 28.9086623 30.52261425 28.8927787