Summarize a fitted model object

Summarize a fitted model object.

Usage

# S3 method for class 'splm'
summary(object, ...)

# S3 method for class 'spautor'
summary(object, ...)

# S3 method for class 'spglm'
summary(object, ...)

# S3 method for class 'spgautor'
summary(object, ...)

Arguments

object

A fitted model object from splm(), spautor(), spglm(), or spgautor().

...

Other arguments. Not used (needed for generic consistency).

Value

A list with several fitted model quantities used to create informative summaries when printing.

Details

summary() creates a summary of a fitted model object intended to be printed using print(). This summary contains useful information like the original function call, residuals, a coefficients table, a pseudo r-squared, and estimated covariance parameters.

See also

Examples

spmod <- splm(z ~ water + tarp,
  data = caribou,
  spcov_type = "exponential", xcoord = x, ycoord = y
)
summary(spmod)
#> 
#> Call:
#> splm(formula = z ~ water + tarp, data = caribou, spcov_type = "exponential", 
#>     xcoord = x, ycoord = y)
#> 
#> Residuals:
#>      Min       1Q   Median       3Q      Max 
#> -0.41281 -0.20763 -0.11205  0.02956  0.45429 
#> 
#> Coefficients (fixed):
#>             Estimate Std. Error z value Pr(>|z|)    
#> (Intercept)  2.04981    0.31093   6.592 4.33e-11 ***
#> waterY      -0.08310    0.06449  -1.289 0.197563    
#> tarpnone     0.08005    0.07759   1.032 0.302166    
#> tarpshade    0.28654    0.07667   3.737 0.000186 ***
#> ---
#> Signif. codes:  0 '***' 0.001 '**' 0.01 '*' 0.05 '.' 0.1 ' ' 1
#> 
#> Pseudo R-squared: 0.3963
#> 
#> Coefficients (exponential spatial covariance):
#>      de      ie   range 
#>  0.1109  0.0226 19.1168