Tidy a fitted model object

Tidy a fitted model object into a summarized tibble.

Usage

# S3 method for splm
tidy(x, conf.int = FALSE, conf.level = 0.95, effects = "fixed", ...)

# S3 method for spautor
tidy(x, conf.int = FALSE, conf.level = 0.95, effects = "fixed", ...)

# S3 method for spglm
tidy(x, conf.int = FALSE, conf.level = 0.95, effects = "fixed", ...)

# S3 method for spgautor
tidy(x, conf.int = FALSE, conf.level = 0.95, effects = "fixed", ...)

Arguments

x

A fitted model object from splm(), spautor(), spglm(), or spgautor().

conf.int

Logical indicating whether or not to include a confidence interval in the tidied output. The default is FALSE.

conf.level

The confidence level to use for the confidence interval if conf.int is TRUE. Must be strictly greater than 0 and less than 1. The default is 0.95, which corresponds to a 95 percent confidence interval.

effects

The type of effects to tidy. Available options are "fixed" (fixed effects), "spcov" (spatial covariance parameters), and "randcov" (random effect variances). The default is "fixed".

...

Other arguments. Not used (needed for generic consistency).

Value

A tidy tibble of summary information effects.

See also

Examples

spmod <- splm(z ~ water + tarp,
  data = caribou,
  spcov_type = "exponential", xcoord = x, ycoord = y
)
tidy(spmod)
#> # A tibble: 4 × 5
#>   term        estimate std.error statistic  p.value
#>   <chr>          <dbl>     <dbl>     <dbl>    <dbl>
#> 1 (Intercept)   2.05      0.311       6.59 4.33e-11
#> 2 waterY       -0.0831    0.0645     -1.29 1.98e- 1
#> 3 tarpnone      0.0801    0.0776      1.03 3.02e- 1
#> 4 tarpshade     0.287     0.0767      3.74 1.86e- 4
tidy(spmod, effects = "spcov")
#> # A tibble: 3 × 3
#>   term  estimate is_known
#>   <chr>    <dbl> <lgl>   
#> 1 de      0.111  FALSE   
#> 2 ie      0.0226 FALSE   
#> 3 range  19.1    FALSE